+++ pytorch_lightning/trainer/connectors/logger_connector.py
@@ -180,7 +180,7 @@ class LoggerConnector:
                     # with a scalar return, auto set it to "val_loss" for callbacks
                     if isinstance(eval_result, torch.Tensor):
                         flat = {'val_loss': eval_result}
-                    else:
+                    elif isinstance(eval_result, dict):
                         flat = flatten_dict(eval_result)
 
                     # removing val_loss magic word to map to checkpoint + ES callback
@@ -461,9 +461,11 @@ class LoggerConnector:
 
         return gathered_epoch_outputs
 
-    def log_train_step_metrics(self, batch_idx, batch_output):
+    def log_train_step_metrics(self, batch_output):
         # when metrics should be logged
-        should_log_metrics = (batch_idx + 1) % self.trainer.row_log_interval == 0 or self.trainer.should_stop
+        should_log_metrics = (
+            (self.trainer.global_step + 1) % self.trainer.row_log_interval == 0 or self.trainer.should_stop
+        )
         if should_log_metrics or self.trainer.fast_dev_run:
             # logs user requested information to logger
             metrics = batch_output.batch_log_metrics